+++ CHANGES.md
@@ -1,5 +1,5 @@
 ### 1.4.0
-* 
+* Added ability to supress key listeners in PopTable.
 
 ### 1.3.1
 * Fixed PopTable from calling layout in an endless loop when automaticallyResized and keepSizedWithinStage are both selected.